fix(mesh): heartbeat is fire-and-forget — stop leaking callbacks that wedge the cache hub
Root cause of the cache-hub wedge (reproduced by a documentation crawl: 163
HeartBeatEvent callbacks pending >30s on the cache hub → RoutingGrain
TimeoutExceptions → /healthz probe times out → pod 0/1 → 503s):

The sync-stream keep-alive did `Post` AND `Observe` the HeartBeatEvent. The
Observe was there to catch a terminal NotFound and stop heart-beating a dead
owner. But a HEALTHY owner never acks a heartbeat, so that Observe registered a
hub callback that never resolved — the Rx `.Timeout` completed the observable
but left the underlying callback pending on the (cache) hub. Across many live
sync streams (e.g. one per open doc page) those leaked callbacks accumulated
until the hub's action block / liveness probe stalled. The heartbeat effectively
waited for an ack it must never wait for.

Fix — make it truly fire-and-forget (the code already claimed to be, and had
already added the change-feed resubscribe as "the sole recycled-grain detector",
but left the leaky Observe in):

- JsonSynchronizationStream: post the heartbeat, do NOT observe it. No callback,
  no leak. Keep-alive still works (the Post reaches live owners); recycled owners
  are re-detected by the existing change-feed resubscribe; a permanently-gone
  owner is heart-beaten into routing's ignore path every interval — a benign
  dropped post, not a storm — until the subscriber hub is itself collected.
- RoutingGrain.PostFailureToSender: skip [CanBeIgnored] messages (HeartBeatEvent,
  Shutdown/Dispose). They are fire-and-forget with no awaiting sender, so a
  DeliveryFailure/NACK for them is meaningless and — for a permanently-gone owner
  heart-beaten every interval — would BE the NotFound storm. This makes the
  Orleans router agree with the monolith RoutingServiceBase, which already
  guarded [CanBeIgnored]. This is the "when not deliverable ⇒ ignore, don't fail
  the delivery" half.

Test: HeartbeatStopsWhenOwnerDiesAfterSubscribeTest pinned the removed
observe-and-stop behaviour; renamed to HeartbeatFireAndForgetTest and inverted —
after the owner dies (NotFound reply), the fire-and-forget heartbeat IGNORES the
reply and keeps flowing (count climbs) instead of freezing. Goes RED if the leaky
Observe is reintroduced. Storm (never-existent owner) + resubscribe tests
unaffected and still pass. Both changed projects build under -warnaserror.
